requirements.txt
psycopg2==2.9.9
python-telegram-bot==20.7
Django==4.2.13